Update dependencies to avoid API breakages.
Some of the libraries the original ACS code depends on have updated
their APIs in a breaking manner.

This commit locks down a few specific libraries to "safe" version ranges
that don't break APIs the code depends on.

Later, it would be nice to port to the newest versions, and avoid all
this hassle.